[Frontned/Fusion] Seperate fusion case
Currently, we support template + activation fusion. But, We found
that there are new available fusion cases which is fusion of
elementwise operations. To do this, we have to manage buffer that
affected by a fusion. There is two type of fusion, vertical or
horizontal fusion. In case of vertical fusion, intermediate buffer
are removed. For horizontal fusion, we avoid loading the buffer
multiple

Plus, set proper tile info for scalar operation